requirements.txt
ansi2html
flake8
mkdocs
mkdocstrings
mkdocs-macros-plugin
mkdocs-material
markdown-include
mypy
pkg/markdown_fortran
python-markdown-math
